Create  LB_Number_Sumeet Sum 1
EXample-1

If the 3 input numbers are 23792, 37221 and 10270.

The largest numbers that can be produced using 3 digits from each of these are 973, 732 and 721 respectively, and the sum of these largest numbers will be 2426.

Therefore the expected result is 2426

Example-2

If the 3 input numbers are 26674, 105 and 37943.

The largest numbers that can be produced using 3 digits from each of these are 766. 510 and 974 respectively. and the sum of these largest numbers will be 2250.

Therefore, the expected result is 2250

Input Format

    Three integers input1, input2, input3

Constraints

All the given 3 numbers will be >=100 and <=99999

Output Format

    The sum of the largest numbers that can be produced using 3 digits